Fix the EmptyArrayAsConst canonicalizer rule
