+++ src/canonicalizer/CMakeLists.txt
@@ -10,6 +10,7 @@ add_library(sourcemeta_jsonbinpack_canonicalizer
   rules/drop_non_boolean_keywords_format.h
   rules/drop_non_boolean_keywords_unevaluated.h
   rules/drop_non_boolean_keywords_validation.h
+  rules/empty_array_as_const.h
   rules/equal_numeric_bounds_as_const.h
   rules/if_without_then_else.h
   rules/implicit_array_lower_bound.h
